employees/views.py
from rest_framework import viewsets
from rest_framework.permissions import IsAuthenticated
from .models import Employee
from .serializers import EmployeeSerializer

class EmployeeViewSet(viewsets.ModelViewSet):
    queryset = Employee.objects.all().order_by("id")
    serializer_class = EmployeeSerializer
    permission_classes = [IsAuthenticated]

    # ✅ Filtering by query params ?department=HR&role=Developer
    filterset_fields = ["department", "role"]
