requirement.txt
Django>=4.2,<6.0
djangorestframework>=3.14,<4.0
djangorestframework-simplejwt>=5.2,<6.0
django-filter>=23.2,<26.0
pytest>=7.4,<9.0
pytest-django>=4.5,<5.0
